chore: convert build_go_test into an integration test
